Added UPDATE operation.

+++ RecursosHumanos/src/co/edu/unbosque/model/persistencia/CandidateDAO.java
@@ -31,40 +31,46 @@ public class CandidateDAO {
 	public boolean addCandidate(String nombre, String apellido, String cargo, String cedula, int edad,
 			ArrayList<CandidateDTO> lst) {
 
-		CandidateDTO agregar = new CandidateDTO(nombre, apellido, cargo, cedula, edad);
+		CandidateDTO agregar = new CandidateDTO(nombre, apellido, cedula, cargo, edad);
 		
-		if (searchCandidate(cedula, lst) == null) {
+		if (searchCandidate(cedula, lst) == null && cedula.matches("[0-9]+") && cedula.length() > 2) {
 			lst.add(agregar);
 			archivo.escribir(lst);
 			return true;
 		} else {
-			JOptionPane.showMessageDialog(null, "El candidato ya se encuentra registrado");
+			JOptionPane.showMessageDialog(null, "El candidato ya se encuentra registrado o la cedula tiene letras.");
 			return false;
 		}
 	}
 	
 	public void deleteCandidate(String cedula, ArrayList<CandidateDTO> lst) {
-		if (!cedula.equals("")) {
-			if (searchCandidate(cedula, lst) != null) {
-				try {
-					CandidateDTO eliminar = searchCandidate(cedula, lst);
-					lst.remove(eliminar);
-					archivo.getArchivo().delete();
-					archivo.getArchivo().createNewFile();
-					archivo.escribir(lst);
-					JOptionPane.showMessageDialog(null, "El candidato se elimino correctamente");
-				} catch (IOException e1) {
-					e1.printStackTrace();
+		
+		if (cedula.matches("[0-9]+") && cedula.length() > 2) {
+			if (!cedula.equals("")) {
+				if (searchCandidate(cedula, lst) != null) {
+					try {
+						CandidateDTO eliminar = searchCandidate(cedula, lst);
+						lst.remove(eliminar);
+						archivo.getArchivo().delete();
+						archivo.getArchivo().createNewFile();
+						archivo.escribir(lst);
+						JOptionPane.showMessageDialog(null, "El candidato se elimino correctamente");
+					} catch (IOException e1) {
+						e1.printStackTrace();
+					}
+				} else {
+					JOptionPane.showMessageDialog(null, "El candidato ingresado no se encuentra registrado");
 				}
-			} else {
-				JOptionPane.showMessageDialog(null, "El candidato ingresado no se encuentra registrado");
 			}
+		} else {
+			JOptionPane.showMessageDialog(null, "La cedula solo debe tener numeros");
 		}
+		
 	}
 	
 	public void updateCandidate (String newname, String newlast, String newid, String newpos, int newage, ArrayList<CandidateDTO> lst) {
 		if (!newid.equals("")) {
-			if (searchCandidate(newid, lst) != null) {
+			if (searchCandidate(newid, lst) != null && newid.matches("[0-9]+")) {
 				try {
 					CandidateDTO toupdate = searchCandidate(newid, lst);
 					toupdate.setName(newname);
@@ -72,13 +78,13 @@ public class CandidateDAO {
 					toupdate.setPosition(newpos);
 					toupdate.setId(newid);
 					toupdate.setAge(newage);
-					archivo.getArchivo().delete();
-					archivo.getArchivo().createNewFile();
-					archivo.escribir(lst);
-				} catch (IOException e1) {
+					JOptionPane.showMessageDialog(null, "Actualizado correctamente");
+				} catch (Exception e1) {
 					e1.printStackTrace();
 				}
 			}
+		} else {
+			JOptionPane.showMessageDialog(null, "Revise el numero de cedula");
 		}
 	}
